+++ app/models/user.rb
@@ -1367,14 +1367,18 @@ class User < Sequel::Model
 
   # INFO: assigning to owner is necessary because of payment reasons
   def assign_geocodings_to_organization_owner
-    return if organization.nil? || organization.owner.nil? || organization_owner?
-    geocodings_dataset.all.each do |g|
-      g.user = organization.owner
-      g.data_import_id = nil
-      g.save(raise_on_failure: true)
+    return if organization&.owner.blank? || organization_owner?
+
+    Carto::Geocoding.where(user_id: id).find_each do |geocoding|
+      geocoding.update!(user_id: organization.owner.id, data_import_id: nil)
     end
   rescue StandardError => e
-    log_error(exception: e, message: 'Error assigning geocodings to org owner', target_user: self)
+    log_error(
+      message: 'Error assigning geocodings to org owner',
+      exception: e,
+      target_user: self,
+      organization: organization
+    )
     geocodings.each(&:destroy)
   end