+++ NEWS.md
@@ -38,6 +38,7 @@ Development
 * Fix KML importing error when the layers have slashes in their names [#15897](https://github.com/CartoDB/cartodb/pull/15897)
 * Create OAuth scope for reading/writing all datasets [#15884](https://github.com/CartoDB/cartodb/pull/15884)
 * Migrate `Organization` model to `ActiveRecord` [#15884](https://github.com/CartoDB/cartodb/pull/15884)
+* Fix bug reassigning geocodings [#15924](https://github.com/CartoDB/cartodb/pull/15924)
 
 4.42.0 (2020-09-28)
 -------------------